@@ -1,10 +1,10 @@
-import React from 'react';
 import clsx from 'clsx';
 import Link from '@docusaurus/Link';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
 import Layout from '@theme/Layout';
 import HomepageFeatures from '@site/src/components/HomepageFeatures';
 
+import Heading from '@theme/Heading';
 import styles from './index.module.css';
 
 function HomepageHeader() {
@@ -12,33 +12,16 @@ function HomepageHeader() {
   return (
     <header className={clsx('hero hero--primary', styles.heroBanner)}>
       <div className="container">
-        <h1 className="hero__title">{siteConfig.title}</h1>
+        <Heading as="h1" className="hero__title">
+          {siteConfig.title}
+        </Heading>
         <p className="hero__subtitle">{siteConfig.tagline}</p>
         <div className={styles.buttons}>
           <Link
             className="button button--secondary button--lg"
-            to="/docs/Tutorial/eggs5">
-            Five minutes
+            to="/docs/intro">
+            Docusaurus Tutorial - 5min ⏱️
           </Link>
-          &nbsp;
-          <Link
-            className="button button--secondary button--lg"
-            to="/docs/Tutorial/eggs-users-guide">
-            Eggs users' guide
-          </Link>
-          &nbsp;
-          <Link
-            className="button button--secondary button--lg"
-            to="/docs/Tutorial/wardrobe-users-guide">
-            Wardrobe users' guide
-          </Link>
-          &nbsp;
-          <Link
-            className="button button--secondary button--lg"
-            to="/docs/faq">
-            FAQ 
-          </Link>
-
         </div>
       </div>
     </header>
@@ -49,8 +32,8 @@ export default function Home() {
   const {siteConfig} = useDocusaurusContext();
   return (
     <Layout
-      title={`${siteConfig.title}`}
-      description="On the road of Remastersys, Refracta, Systemback and father Knoppix!">
+      title={`Hello from ${siteConfig.title}`}
+      description="Description will go into a meta tag in <head />">
       <HomepageHeader />
       <main>
         <HomepageFeatures />
